Add support for reviewer changes in merge requests
The merge request event webhook is called for reviwers updates, however
this is ignored by the GitlabChanges class. Therefore these updates are
not forwarded to Matrix. By adding the reviewers section to both the
GitlabChanges and to the issue_update template (which is used for merge
request updates) the bot reports about changes in reviewers.